requirements.txt
requests
tdqm
matplotlib
python-dotenv
langchain
langchain-mistralai
langchain-openai
langchain-community
langchain-chroma
langgraph